requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.55
botocore==1.31.55
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
cssutils==2.9.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.57.0
django-csp==3.7
django-hosts==5.2
django-inlinecss==0.3.0
django-phonenumber-field==7.1.0
django-storages==1.14.1
et-xmlfile==1.1.0
frozenlist==1.4.0
future==0.18.3
gunicorn==21.2.0
httpie==3.2.2
idna==3.4
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
naomi==0.1.2
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
phonenumbers==8.13.21
Pillow==10.1.0
psycopg==3.1.12
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pynliner==0.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
s3transfer==0.7.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
twilio==8.9.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
yarl==1.9.2
